runtime: match go1.26 builtin print float format
