+++ runtime/internal/runtime/z_print.go
@@ -37,48 +37,6 @@ func PrintByte(v byte) {
 	c.Fputc(c.Int(v), c.Stderr)
 }
 
-func PrintFloat(v float64) {
-	switch {
-	case v != v:
-		c.Fprintf(c.Stderr, c.Str("NaN"))
-		return
-	case v+v == v && v != 0:
-		if v > 0 {
-			c.Fprintf(c.Stderr, c.Str("+Inf"))
-		} else {
-			c.Fprintf(c.Stderr, c.Str("-Inf"))
-		}
-		return
-	}
-	// Go's builtin print/println formats float exponents with at least 3 digits,
-	// while libc "%+e" typically uses 2. Normalize the exponent width so runtime
-	// output matches the standard toolchain in goroot run cases.
-	buf := (*c.Char)(c.Alloca(32))
-	c.Snprintf(buf, 32, c.Str("%+e"), v)
-	n := c.Strlen(buf)
-	if n >= 4 {
-		exp := n - 4
-		ch := c.Index(buf, exp)
-		sign := c.Index(buf, exp+1)
-		d0 := c.Index(buf, exp+2)
-		d1 := c.Index(buf, exp+3)
-		if (ch == 'e' || ch == 'E') &&
-			(sign == '+' || sign == '-') &&
-			d0 >= '0' && d0 <= '9' &&
-			d1 >= '0' && d1 <= '9' {
-			dst := unsafe.Pointer(c.Advance(buf, exp+3))
-			src := unsafe.Pointer(c.Advance(buf, exp+2))
-			c.Memmove(dst, src, 3)
-			*c.Advance(buf, exp+2) = '0'
-		}
-	}
-	c.Fputs(buf, c.Stderr)
-}
-
-func PrintComplex(v complex128) {
-	print("(", real(v), imag(v), "i)")
-}
-
 func PrintUint(v uint64) {
 	c.Fprintf(c.Stderr, printFormatPrefixUInt, v)
 }
